Moved view functionality to Codebehind instead of Main
