+++ Code/src/.classpath
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="src" path="main/java"/>
+	<classpathentry kind="src" path="test/java"/>
+	<classpathentry kind="con" path="org.eclipse.fx.ide.jdt.core.JAVAFX_CONTAINER"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx-swt.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.base.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.controls.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.fxml.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.graphics.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.media.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.swing.jar"/>
+	<classpathentry kind="var" path="ECLIPSE_HOME/javafx-sdk-11.0.2/lib/javafx.web.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
